+++ R/missing.R
@@ -1,0 +1,5 @@
+#' @method is.na vecvec
+#' @export
+is.na.vecvec <- function(x) {
+  unvecvec(vecvec_apply(x, is.na))
+}